+++ whitelabel/docker/start-whitelabel-services.sh
@@ -1,3 +1,3 @@
 #!/usr/bin/env bash
 
-./start-services.sh -i both --with-hbase --with-hdfs $@
+./start-services.sh -i both --with-hbase --with-hdfs --with-mysql $@