+++ RELEASE NOTES.md
@@ -140,4 +140,6 @@
 ...
 
 **Update**
-1- Modificato chroot_path di zookeeper per Kafka: da "" a "/kafka" (allineamento con Solr per cui è "/solr")
+- Modificato `chroot_path` di zookeeper per Kafka: da "" a "/kafka" (allineamento con Solr per cui è "/solr")
+
+- JDBCReader plugin